refactored packaging conventions
